toolbars: Перенес утилиты в библиотеку https://online.sbis.ru/opendoc.html?guid=d00e9811-b1ba-4570-be41-b20a48d151ff
