+++ Controls-demo/grid/EditInPlace/Toolbar/Index.ts
@@ -6,7 +6,7 @@ import {Memory} from 'Types/source';
 import {Model} from 'Types/entity';
 import {getCountriesStats} from '../../DemoHelpers/DataCatalog';
 import {SyntheticEvent} from 'Vdom/Vdom';
-import {showType} from 'Controls/Utils/Toolbar';
+import {showType} from 'Controls/toolbars';
 import { IColumn } from 'Controls/grid';
 import {IItemAction} from 'Controls/itemActions';
