+++ Controls/Utils/Toolbar.ts
@@ -58,44 +58,4 @@
  * @public
  */
 
-import {Abstract as ChainAbstract, factory} from 'Types/chain';
-import {RecordSet} from 'Types/collection';
-import {Record} from 'Types/entity';
-
-export interface IShowType {
-   /**
-    * Show only in Menu.
-    */
-   MENU: number;
-   /**
-    * Show in Menu and Toolbar.
-    */
-   MENU_TOOLBAR: number;
-   /**
-    * Show only in Toolbar
-    */
-   TOOLBAR: number;
-}
-
-export enum showType {
-   MENU,
-   MENU_TOOLBAR,
-   TOOLBAR
-}
-
-export function getMenuItems<T extends Record>(items: RecordSet<T> | T[]): ChainAbstract<T> {
-   return factory(items).filter((item) => {
-      return item.get('showType') !== this.showType.TOOLBAR;
-   });
-}
-
-export function needShowMenu(items: RecordSet): boolean {
-    const enumerator = items.getEnumerator();
-    while (enumerator.moveNext()) {
-        if (enumerator.getCurrent().get('showType') !== this.showType.TOOLBAR) {
-            return true;
-        }
-    }
-
-    return false;
-}
+export {showType, IShowType, getMenuItems, needShowMenu} from 'Controls/toolbars';
