+++ Controls/_operationsPanel/OperationsPanel/Utils.ts
@@ -1,6 +1,11 @@
-import toolbars = require('Controls/toolbars');
-import {showType} from 'Controls/Utils/Toolbar';
-import { Logger } from 'UI/Utils';
+import {
+   actualItems,
+   getButtonTemplate,
+   getButtonTemplateOptionsByItem,
+   ItemTemplate,
+   showType
+} from 'Controls/toolbars';
+import {Logger} from 'UI/Utils';
 import {Record} from 'Types/entity';
 import {DOMUtil, getWidth} from 'Controls/sizeUtils';
 import {constants} from 'Env/Env';
@@ -19,7 +24,7 @@ import {constants} from 'Env/Env';
          if (itemTemplateProperty && item) {
             contentTemplate = item.get(itemTemplateProperty);
          }
-         if (!contentTemplate && itemTemplate !== toolbars.ItemTemplate) {
+         if (!contentTemplate && itemTemplate !== ItemTemplate) {
             contentTemplate = itemTemplate;
          }
          return contentTemplate;
@@ -34,12 +39,12 @@ import {constants} from 'Env/Env';
             item = items.getRecordById(key);
             buttonTemplateOptions = _private.getButtonTemplateOptionsForItem(item, itemTemplateProperty);
 
-            itemsMark.push(toolbars.ItemTemplate({
+            itemsMark.push(ItemTemplate({
                item,
                size: 'm',
                itemsSpacing: 'medium',
                theme,
-               buttonTemplate: toolbars.getButtonTemplate(),
+               buttonTemplate: getButtonTemplate(),
                buttonTemplateOptions,
                contentTemplate: _private.getContentTemplate(item, itemTemplate, itemTemplateProperty)
             }));
@@ -49,7 +54,7 @@ import {constants} from 'Env/Env';
       },
 
       getButtonTemplateOptionsForItem(item: Record, itemTemplateProperty?: string): object {
-         const buttonOptions = toolbars.getButtonTemplateOptionsByItem(item);
+         const buttonOptions = getButtonTemplateOptionsByItem(item);
 
          if (itemTemplateProperty &&
              item.get(itemTemplateProperty) &&
@@ -77,7 +82,7 @@ import {constants} from 'Env/Env';
             currentWidth,
             visibleItemsKeys = [];
 
-         toolbars.actualItems(items);
+         actualItems(items);
 
          items.each(function(item) {
             if (!item.get(parentProperty)) {
