+++ Controls/toolbars.ts
@@ -11,6 +11,7 @@ export {ItemTemplate};
 
 export {items as actualItems} from 'Controls/_toolbars/ActualAPI';
 export {default as View, IToolbarOptions, TItemsSpacing} from './_toolbars/View';
-export {getButtonTemplateOptionsByItem, getButtonTemplate} from './_toolbars/Util';
+export {getButtonTemplateOptionsByItem, getButtonTemplate, getMenuItems, needShowMenu} from './_toolbars/Util';
 export {default as BoxView} from './_toolbars/BoxView';
 export {default as IToolbarSource} from './_toolbars/IToolbarSource';
+export {showType, IShowType} from './_toolbars/interfaces/IShowType';
